DDP-5758 - remove services and other small classes (join into bigger ones) - try to make structure of AI builder module more clear
