+++ pepper-apis/src/main/java/org/broadinstitute/ddp/service/actvityinstancebuilder/form/block/FormBlockCreator.java
@@ -22,7 +22,7 @@ public class FormBlockCreator {
     }
 
     private FormBlock constructFormBlock(AIBuilderContext ctx, FormBlockDef formBlockDef) {
-        FormBlockCreatorHelper creatorHelper = ctx.getAIBuilderFactory().getAICreatorsFactory().getFormBlockCreatorHelper();
+        FormBlockCreatorHelper creatorHelper = ctx.getAIBuilderFactory().getFormBlockCreatorHelper();
         switch (formBlockDef.getBlockType()) {
             case GROUP:
                 return creatorHelper.createGroupBlock(ctx, (GroupBlockDef) formBlockDef);
