+++ pepper-apis/src/main/java/org/broadinstitute/ddp/service/actvityinstancebuilder/form/block/question/QuestionCreatorHelper.java
@@ -31,66 +31,66 @@ public class QuestionCreatorHelper {
     AgreementQuestion createAgreementQuestion(AIBuilderContext ctx, AgreementQuestionDef questionDef) {
         return new AgreementQuestion(
                 questionDef.getStableId(),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getPromptTemplate()),
                 questionDef.isRestricted(),
                 questionDef.isDeprecated(),
                 isReadOnly(questionDef, ctx.getFormResponse().getLatestStatus().getType(), ctx.getPreviousInstanceId()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getTooltipTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoHeaderTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoFooterTemplate()),
-                ctx.getAIBuilderFactory().getAICreatorsFactory().getQuestionCreator().getAnswers(ctx, questionDef.getStableId()),
-                ctx.getAIBuilderFactory().getAICreatorsFactory().getQuestionCreator().getValidationRules(ctx, questionDef)
+                ctx.getAIBuilderFactory().getQuestionCreator().getAnswers(ctx, questionDef.getStableId()),
+                ctx.getAIBuilderFactory().getQuestionCreator().getValidationRules(ctx, questionDef)
         );
     }
 
     BoolQuestion createBoolQuestion(AIBuilderContext ctx, BoolQuestionDef questionDef) {
         return new BoolQuestion(
                 questionDef.getStableId(),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getPromptTemplate()),
                 questionDef.isRestricted(),
                 questionDef.isDeprecated(),
                 isReadOnly(questionDef, ctx.getFormResponse().getLatestStatus().getType(), ctx.getPreviousInstanceId()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getTooltipTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoHeaderTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoFooterTemplate()),
-                ctx.getAIBuilderFactory().getAICreatorsFactory().getQuestionCreator().getAnswers(ctx, questionDef.getStableId()),
-                ctx.getAIBuilderFactory().getAICreatorsFactory().getQuestionCreator().getValidationRules(ctx, questionDef),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getQuestionCreator().getAnswers(ctx, questionDef.getStableId()),
+                ctx.getAIBuilderFactory().getQuestionCreator().getValidationRules(ctx, questionDef),
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getTrueTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getFalseTemplate())
         );
     }
 
     CompositeQuestion createCompositeQuestion(AIBuilderContext ctx, CompositeQuestionDef questionDef) {
-        QuestionCreator questionCreator = ctx.getAIBuilderFactory().getAICreatorsFactory().getQuestionCreator();
+        QuestionCreator questionCreator = ctx.getAIBuilderFactory().getQuestionCreator();
         return new CompositeQuestion(
                 questionDef.getStableId(),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getPromptTemplate()),
                 questionDef.isRestricted(),
                 questionDef.isDeprecated(),
                 isReadOnly(questionDef, ctx.getFormResponse().getLatestStatus().getType(), ctx.getPreviousInstanceId()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getTooltipTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoHeaderTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoFooterTemplate()),
                 questionCreator.getValidationRules(ctx, questionDef),
                 questionDef.isAllowMultiple(),
                 questionDef.isUnwrapOnExport(),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAddButtonTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalItemTemplate()),
                 CollectionMiscUtil.createListFromAnotherList(questionDef.getChildren(),
                         (childQuestionDef) -> questionCreator.createQuestion(ctx, childQuestionDef)),
@@ -100,26 +100,26 @@ public class QuestionCreatorHelper {
     }
 
     DatePicklistQuestion createDatePickListQuestion(AIBuilderContext ctx, DateQuestionDef questionDef) {
-        QuestionCreator questionCreator = ctx.getAIBuilderFactory().getAICreatorsFactory().getQuestionCreator();
+        QuestionCreator questionCreator = ctx.getAIBuilderFactory().getQuestionCreator();
         return new DatePicklistQuestion(
                 questionDef.getStableId(),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getPromptTemplate()),
                 questionDef.isRestricted(),
                 questionDef.isDeprecated(),
                 isReadOnly(questionDef, ctx.getFormResponse().getLatestStatus().getType(), ctx.getPreviousInstanceId()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getTooltipTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoHeaderTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoFooterTemplate()),
                 questionCreator.getAnswers(ctx, questionDef.getStableId()),
                 questionCreator.getValidationRules(ctx, questionDef),
                 questionDef.getRenderMode(),
                 questionDef.isDisplayCalendar(),
                 questionDef.getFields(),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getPlaceholderTemplate()),
                 questionDef.getPicklistDef().getUseMonthNames(),
                 questionDef.getPicklistDef().getStartYear(),
@@ -129,44 +129,44 @@ public class QuestionCreatorHelper {
     }
 
     DateQuestion createDateQuestion(AIBuilderContext ctx, DateQuestionDef questionDef) {
-        QuestionCreator questionCreator = ctx.getAIBuilderFactory().getAICreatorsFactory().getQuestionCreator();
+        QuestionCreator questionCreator = ctx.getAIBuilderFactory().getQuestionCreator();
         return new DateQuestion(
                 questionDef.getStableId(),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getPromptTemplate()),
                 questionDef.isRestricted(),
                 questionDef.isDeprecated(),
                 isReadOnly(questionDef, ctx.getFormResponse().getLatestStatus().getType(), ctx.getPreviousInstanceId()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getTooltipTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoHeaderTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoFooterTemplate()),
                 questionCreator.getAnswers(ctx, questionDef.getStableId()),
                 questionCreator.getValidationRules(ctx, questionDef),
                 questionDef.getRenderMode(),
                 questionDef.isDisplayCalendar(),
                 questionDef.getFields(),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getPlaceholderTemplate())
         );
     }
 
     FileQuestion constructFileQuestion(AIBuilderContext ctx, FileQuestionDef questionDef) {
-        QuestionCreator questionCreator = ctx.getAIBuilderFactory().getAICreatorsFactory().getQuestionCreator();
+        QuestionCreator questionCreator = ctx.getAIBuilderFactory().getQuestionCreator();
         return new FileQuestion(
                 questionDef.getStableId(),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getPromptTemplate()),
                 questionDef.isRestricted(),
                 questionDef.isDeprecated(),
                 isReadOnly(questionDef, ctx.getFormResponse().getLatestStatus().getType(), ctx.getPreviousInstanceId()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getTooltipTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoHeaderTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoFooterTemplate()),
                 questionCreator.getAnswers(ctx, questionDef.getStableId()),
                 questionCreator.getValidationRules(ctx, questionDef)
@@ -174,21 +174,21 @@ public class QuestionCreatorHelper {
     }
 
     NumericQuestion createNumericQuestion(AIBuilderContext ctx, NumericQuestionDef questionDef) {
-        QuestionCreator questionCreator = ctx.getAIBuilderFactory().getAICreatorsFactory().getQuestionCreator();
+        QuestionCreator questionCreator = ctx.getAIBuilderFactory().getQuestionCreator();
         return new NumericQuestion(
                 questionDef.getStableId(),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getPromptTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getPlaceholderTemplate()),
                 questionDef.isRestricted(),
                 questionDef.isDeprecated(),
                 isReadOnly(questionDef, ctx.getFormResponse().getLatestStatus().getType(), ctx.getPreviousInstanceId()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getTooltipTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoHeaderTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoFooterTemplate()),
                 questionCreator.getAnswers(ctx, questionDef.getStableId()),
                 questionCreator.getValidationRules(ctx, questionDef),
@@ -197,35 +197,35 @@ public class QuestionCreatorHelper {
     }
 
     PicklistQuestion createPicklistQuestion(AIBuilderContext ctx, PicklistQuestionDef questionDef) {
-        QuestionCreator questionCreator = ctx.getAIBuilderFactory().getAICreatorsFactory().getQuestionCreator();
+        QuestionCreator questionCreator = ctx.getAIBuilderFactory().getQuestionCreator();
 
         List<PicklistGroup> picklistGroups = CollectionMiscUtil.createListFromAnotherList(questionDef.getGroups(),
-                (picklistGroupDef) -> ctx.getAIBuilderFactory().getAICreatorsFactory().getPicklistCreatorHelper()
+                (picklistGroupDef) -> ctx.getAIBuilderFactory().getPicklistCreatorHelper()
                         .createPicklistGroup(ctx, picklistGroupDef));
 
         List<PicklistOption> picklistOptions = CollectionMiscUtil.createListFromAnotherList(questionDef.getAllPicklistOptions(),
                 (picklistOptionDef) ->
-                        ctx.getAIBuilderFactory().getAICreatorsFactory().getPicklistCreatorHelper()
+                        ctx.getAIBuilderFactory().getPicklistCreatorHelper()
                                 .createPicklistOption(ctx, picklistOptionDef, questionDef.getGroups()));
 
         return new PicklistQuestion(
                 questionDef.getStableId(),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getPromptTemplate()),
                 questionDef.isRestricted(),
                 questionDef.isDeprecated(),
                 isReadOnly(questionDef, ctx.getFormResponse().getLatestStatus().getType(), ctx.getPreviousInstanceId()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getTooltipTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoHeaderTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoFooterTemplate()),
                 questionCreator.getAnswers(ctx, questionDef.getStableId()),
                 questionCreator.getValidationRules(ctx, questionDef),
                 questionDef.getSelectMode(),
                 questionDef.getRenderMode(),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getPicklistLabelTemplate()),
                 picklistOptions,
                 picklistGroups
@@ -233,21 +233,21 @@ public class QuestionCreatorHelper {
     }
 
     TextQuestion createTextQuestion(AIBuilderContext ctx, TextQuestionDef questionDef) {
-        QuestionCreator questionCreator = ctx.getAIBuilderFactory().getAICreatorsFactory().getQuestionCreator();
+        QuestionCreator questionCreator = ctx.getAIBuilderFactory().getQuestionCreator();
         return new TextQuestion(
                 questionDef.getStableId(),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getPromptTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getPlaceholderTemplate()),
                 questionDef.isRestricted(),
                 questionDef.isDeprecated(),
                 isReadOnly(questionDef, ctx.getFormResponse().getLatestStatus().getType(), ctx.getPreviousInstanceId()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getTooltipTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoHeaderTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getAdditionalInfoFooterTemplate()),
                 questionCreator.getAnswers(ctx, questionDef.getStableId()),
                 questionCreator.getValidationRules(ctx, questionDef),
@@ -255,9 +255,9 @@ public class QuestionCreatorHelper {
                 questionDef.getSuggestionType(),
                 questionDef.getSuggestions(),
                 questionDef.isConfirmEntry(),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getConfirmPromptTemplate()),
-                ctx.getAIBuilderFactory().getTemplateRenderFactory().renderTemplate(
+                ctx.getAIBuilderFactory().getTemplateRenderHelper().renderTemplate(
                         ctx, questionDef.getMismatchMessageTemplate())
         );
     }
